Move config module to root.
